requirements.txt
# 核心依赖
xmindparser>=1.0.0
xmind-sdk-python>=1.0.0

# AI 集成
openai>=1.0.0  # 或使用其他 LLM SDK

# 文档解析
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0

# SQL 生成
jinja2>=3.1.0

# CLI
click>=8.1.0

# 测试
pytest>=7.0.0
